refactor(-): :hammer: cria subapp solicitacao_dieta_especial; move models
