+++ sme_sigpae_api/dieta_especial/__tests__/commands/test_verificar_solicitacoes_sem_protocolo.py
@@ -7,7 +7,9 @@ from django.core.management import call_command
 from sme_sigpae_api.dieta_especial.fixtures.factories.dieta_especial_base_factory import (
     SolicitacaoDietaEspecialFactory,
 )
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 
 
 @pytest.mark.django_db
