+++ sme_sigpae_api/dieta_especial/__tests__/logs_models/test_tasks.py
@@ -7,7 +7,9 @@ from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadas,
     LogQuantidadeDietasAutorizadasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 from sme_sigpae_api.dieta_especial.tasks.logs import (
     gera_logs_dietas_especiais_diariamente,
 )
