+++ sme_sigpae_api/dieta_especial/__tests__/tasks/processamentos/test_processa_dietas_especiais_task.py
@@ -12,7 +12,9 @@ from sme_sigpae_api.dados_comuns.models import LogSolicitacoesUsuario
 from sme_sigpae_api.dieta_especial.fixtures.factories.dieta_especial_base_factory import (
     SolicitacaoDietaEspecialFactory,
 )
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 from sme_sigpae_api.dieta_especial.tasks import (
     cancela_dietas_ativas_automaticamente_task,
     cancela_dietas_pendente_autorizacao_task,
