+++ sme_sigpae_api/dieta_especial/__tests__/tasks/test_setup_relatorio_recreio.py
@@ -4,7 +4,9 @@ from sme_sigpae_api.dieta_especial.fixtures.factories.dieta_especial_base_factor
     ClassificacaoDietaFactory,
     SolicitacaoDietaEspecialFactory,
 )
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 from sme_sigpae_api.escola.fixtures.factories.escola_factory import (
     AlunoFactory,
     DiretoriaRegionalFactory,
