+++ sme_sigpae_api/dieta_especial/management/commands/dieta_especial_cancelar.py
@@ -1,7 +1,9 @@
 from django.core.management.base import BaseCommand
 from faker import Faker
 
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 from sme_sigpae_api.perfil.models import Usuario
 
 faker = Faker()
