+++ sme_sigpae_api/dieta_especial/management/commands/verificar_solicitacoes_sem_protocolo.py
@@ -5,7 +5,9 @@ from django.conf import settings
 from django.core.management.base import BaseCommand
 from openpyxl import Workbook
 
-from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    SolicitacaoDietaEspecial,
+)
 
 
 class Command(BaseCommand):
