+++ sme_sigpae_api/dieta_especial/tasks/relatorio_terceirizadas_xlsx.py
@@ -29,7 +29,9 @@ logger = logging.getLogger(__name__)
 def gera_xlsx_relatorio_dietas_especiais_terceirizadas_async(
     user, nome_arquivo, ids_dietas, data, lotes, classificacoes, protocolos_padrao
 ):
-    from sme_sigpae_api.dieta_especial.models import SolicitacaoDietaEspecial
+    from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+        SolicitacaoDietaEspecial,
+    )
 
     logger.info(f"x-x-x-x Iniciando a geração do arquivo {nome_arquivo} x-x-x-x")
     obj_central_download = gera_objeto_na_central_download(
@@ -49,8 +51,12 @@ def gera_xlsx_relatorio_dietas_especiais_terceirizadas_async(
                 "nome_escola": solicitacao.escola.nome,
                 "classificacao": classificacao,
                 "protocolo_padrao": (
-                    solicitacao.nome_protocolo if solicitacao.nome_protocolo else (
-                        solicitacao.protocolo_padrao.nome_protocolo if solicitacao.protocolo_padrao else "--"
+                    solicitacao.nome_protocolo
+                    if solicitacao.nome_protocolo
+                    else (
+                        solicitacao.protocolo_padrao.nome_protocolo
+                        if solicitacao.protocolo_padrao
+                        else "--"
                     )
                 ),
                 "alergias_intolerancias": solicitacao.alergias_intolerancias,
