+++ utility/carga_dados/cardapio/_3_solicitacoes_dieta_especial.py
@@ -11,11 +11,13 @@ from faker import Faker
 from sme_sigpae_api.dieta_especial.models import (
     AlergiaIntolerancia,
     Alimento,
-    Anexo,
     ClassificacaoDieta,
     MotivoNegacao,
-    SolicitacaoDietaEspecial,
     SubstituicaoAlimento,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
+    Anexo,
+    SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.escola.models import Aluno, Escola
 from sme_sigpae_api.perfil.models import Usuario
